fix selected absolute state and move thresholdInput to its own file
